move some long-running tests into a separate module for nightly tests
